Fix execution bugs and improve UI

+++ src/main/java/service/InscricaoService.java
@@ -10,8 +10,6 @@ import repository.UsuarioRepository;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
 
 public class InscricaoService {
     InscricaoRepository inscricaoRepository;
@@ -24,29 +22,39 @@ public class InscricaoService {
     }
 
     public Boolean realizarInscricao(Evento evento, Usuario usuario) throws Exception{
-        if(evento.getCapacidadeAtual().equals(evento.getCapacidadeMax())) throw new Exception("Capacidade máxima atingida.");
+        if(evento.getCapacidadeAtual() >= (evento.getCapacidadeMax())) throw new Exception("Capacidade máxima atingida.");
+        var eventoJaInscrito = usuario.getInscricoes().stream()
+                .filter(i -> i.getEvento().getIdEvento().equals(evento.getIdEvento()))
+                .findFirst();
+        if(eventoJaInscrito.isPresent()) throw new Exception("Você já se inscreveu para esse evento");
         var data = LocalDate.now();
         if(data.isAfter(evento.getDataInicio())) throw new Exception("Prazo para inscrição já acabou.");
-        inscricaoRepository.create(new Inscricao(
+        Inscricao inscricao = new Inscricao(
                 data,
                 "INSCRITO",
                 evento,
                 usuario
-        ));
+        );
         evento.setCapacidadeAtual(evento.getCapacidadeAtual()+1);
+        evento.getInscricoes().add(inscricao);
+        usuario.getInscricoes().add(inscricao);
+        inscricaoRepository.create(inscricao);
+        usuarioRepository.update(usuario);
+        eventoRepository.update(evento);
         return true;
     }
 
     public Boolean cancelarInscricao(Evento evento, Usuario usuario) throws Exception{
         var inscricaoOptional = inscricaoRepository.getByUserIdAndEventId(evento.getIdEvento(), usuario.getIdUsuario());
         if(inscricaoOptional.isEmpty()) throw new Exception("Você não está inscrito para o evento: " + evento.getNome());
+        if(LocalDate.now().isAfter(evento.getDataInicio().minusDays(evento.getDiasCancelarInscricao()))) throw new Exception("Prazo para cancelar inscrição já encerrou.");
         var inscricaoFound = inscricaoOptional.get();
-        inscricaoRepository.delete(inscricaoFound.getIdInscricao());
         evento.getInscricoes().removeIf(i -> i.getIdInscricao().equals(inscricaoFound.getIdInscricao()));
-        eventoRepository.update(evento);
+        evento.setCapacidadeAtual(evento.getInscricoes().size()-1);
         usuario.getInscricoes().removeIf(i -> i.getIdInscricao().equals(inscricaoFound.getIdInscricao()));
+        eventoRepository.update(evento);
         usuarioRepository.update(usuario);
-        return true;
+        return inscricaoRepository.delete(inscricaoFound.getIdInscricao());
     }
 
     public ArrayList<Inscricao> visualizarInscritosNoEvento(Evento evento){
